refactor: name the invoice confirmation modals and split the status helpers
Void, reset, and write-off render BaseConfirmationModal, so they take the
ConfirmationModal suffix alongside mileage and time tracking. The discard
modal becomes InvoiceDiscardChangesConfirmationModal so it reads entity-first
like its siblings. InvoiceRefundModal keeps its name: it hosts
InvoiceRefundForm with closeOnConfirm disabled, so it is a form modal wearing
the confirmation shell rather than a confirmation.

statusComponents.tsx was a loose .tsx at the domain root holding both a pure
status descriptor and a JSX helper. The descriptor moves into the domain
utils as getInvoiceStatusDisplay, and the helper becomes an
InvoiceStatusBadge component with its own directory.
