+++ src/components/features/invoices/InvoiceDiscardChangesConfirmationModal/InvoiceDiscardChangesConfirmationModal.tsx
@@ -2,17 +2,17 @@ import { useTranslation } from 'react-i18next'
 
 import { BaseConfirmationModal } from '@blocks/BaseConfirmationModal/BaseConfirmationModal'
 
-export type DiscardInvoiceChangesConfirmationModalProps = {
+export type InvoiceDiscardChangesConfirmationModalProps = {
   isOpen: boolean
   onOpenChange: (isOpen: boolean) => void
   onConfirm: () => void
 }
 
-export const DiscardInvoiceChangesConfirmationModal = ({
+export const InvoiceDiscardChangesConfirmationModal = ({
   isOpen,
   onOpenChange,
   onConfirm,
-}: DiscardInvoiceChangesConfirmationModalProps) => {
+}: InvoiceDiscardChangesConfirmationModalProps) => {
   const { t } = useTranslation()
   return (
     <BaseConfirmationModal
