+++ src/components/features/invoices/InvoicesMobileList/InvoicesMobileListItemStatusFooter.tsx
@@ -3,13 +3,13 @@ import { useTranslation } from 'react-i18next'
 import { type Invoice } from '@schemas/invoices/invoice'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { MobileListItemStatusFooter } from '@ui/MobileList/MobileListItemStatusFooter'
-import { getInvoiceStatusComponents } from '@features/invoices/statusComponents'
+import { getInvoiceStatusDisplay } from '@features/invoices/utils'
 
 export const InvoicesMobileListItemStatusFooter = ({ invoice }: { invoice: Invoice }) => {
   const { t } = useTranslation()
   const { formatNumber } = useIntlFormatter()
 
-  const { variant, text, subText, Icon } = getInvoiceStatusComponents(invoice, t, formatNumber)
+  const { variant, text, subText, Icon } = getInvoiceStatusDisplay(invoice, t, formatNumber)
 
   return (
     <MobileListItemStatusFooter
