--- src/components/features/profitAndLoss/BaseSummariesBreakdownFooter/BaseSummariesBreakdownFooterContainer.tsx
@@ -1,7 +1,0 @@
-import { type PropsWithChildren } from 'react'
-
-import './baseSummariesBreakdownFooterContainer.scss'
-
-export function BaseSummariesBreakdownFooterContainer({ children }: PropsWithChildren) {
-  return <div className='Layer__BaseSummariesBreakdownFooter'>{children}</div>
-}
