+++ src/components/features/profitAndLoss/ProfitAndLossSummaryTileFooter/profitAndLossSummaryTileFooter.scss
