+++ src/components/features/profitAndLoss/ProfitAndLossSummaryTileFooter/profitAndLossSummaryTileFooterRow.scss
